Add commands to run examples and record results
Add commands to extract examples, run them, and record results. The
data is stored in an extra JSON file.

A configuration file ~/.rpc_docs_helper.yaml is used to specify how
clients are called to run the example RPC commands. It's possible to
configure a bitcoin-cli client or to use curl to access a node via
HTTP. Which client is used can be defined per command.

Examples can be extracted from the help output of a client with the
`update-examples` sub-command.

Results of commands run against a regtest or real node are can be
recorded with the sub-command `update-example-results`.

A single example can be executed with the command `run-example`.

To run a raw command against the configured client you can use the
command `run-cmd`.

Results depend quite a bit on the state of the client. So not all
commands succeed and results are not necessarily reproducible.

The commit also removes the BitcoinCli class in favor of more direct
and simple handling of command line calls through CliCaller.