Add examples output
Add file examples-bitcoin-0.18.json which contains extracted examples
and results of running the example against a node. Results are mostly
from a regtest node plus some results from a mainnet node. The
examples depend on the state of the node and are not all valid in
itself, so they have to be taken within this context.