Add simple web UI for executing RPC commands
The `ui` script starts a web server which provides a simple UI for
executing RPC commands. It requires a running Bitcoin Core node and
a cofigured command line client for executing RPC commands.

The UI automatically populates the list of commands, their help, and
the forms for providing arguments and executing them based on the
output of the RPC help of the running client.

The UI needs a bit of visual polishing and some more thorough handling
of error and corner cases, but it is mostly functional at this stage.